requirements.txt
customtkinter==5.2.2
Pillow>=9.5.0,<10.0.0
psutil==5.9.6
matplotlib==3.8.2
send2trash==1.8.2
watchdog==3.0.0
PTable==0.9.2